Some modification for the report generation process and some troubleshooting information

+++ generate_pdf_report.py
@@ -12,6 +12,9 @@ parser.add_argument("--results_dir", type=str, required=True,  help="""The direc
 parser.add_argument("--title", type=str, required=True, help="""Title of the html and pdf file.""")
 
 parser.add_argument("--output_dir", type=str, default="./", help="""The directory where the html and pdf file will be stored. Default is current directory""")
+parser.add_argument("--keep_html", action="store_true", default=False, 
+    help="""Whether keep the html reports from which pdf reports can be built. 
+    This is useful if the conversion between html and pdf fails.""")
 
 args = parser.parse_args()
 
@@ -24,8 +27,8 @@ output_dir = os.path.join('./', 'report_files')
 if not os.path.exists(output_dir):
     os.makedirs(output_dir)
 
-output_html = os.path.join(output_dir, 'report -- {}.html'.format(args.title))
-output_pdf = os.path.join(output_dir, 'report -- {}.pdf'.format(args.title))
+output_html = os.path.join(output_dir, 'report_{}.html'.format(args.title))
+output_pdf = os.path.join(output_dir, 'report_{}.pdf'.format(args.title))
 
 images_path = os.path.abspath(os.path.join(os.path.dirname( __file__ ), args.results_dir))
 
@@ -152,6 +155,15 @@ for super_cat in super_categories:
     
     
     episodes.append(super_cat_dic)
+
+# calculate the average AUC across episodes
+average_AUC = 0
+cnt = 0
+for episode in episodes:
+    average_AUC += float(episode["AUC"])
+    cnt += 1
+average_AUC /= cnt
+average_AUC = "{:.2f}".format(average_AUC)
     
     
 subs = jinja2.Environment(
@@ -161,7 +173,8 @@ subs = jinja2.Environment(
                                        total_categories=total_categories,
                                        categories_combined=categories_combined,
                                        over_all_auc_histogram=over_all_auc_histogram,
-                                       episodes=episodes)
+                                       episodes=episodes,
+                                       average_AUC=average_AUC)
 
 
 # lets write the substitution to a file
@@ -187,7 +200,8 @@ print("PDF File is Ready")
 print("###########################")
 print()
 
-os.remove(output_html)
+if not args.keep_html:
+    os.remove(output_html)
